refactor(db): extract cartog-db test module into focused files
The 4713-line #[cfg(test)] mod tests (194 tests) moves out of lib.rs into
src/tests/{normalize,crud,resolution,traversal,rag,migration,errors,
atomicity,resolution_state}.rs. lib.rs drops 5809 -> 1098 lines; production
code byte-identical. Tests stay in-crate (private access via crate::*);
test_symbol hoisted to the tests hub. No test added, dropped, or modified.
